// src/types/project.ts
export interface Project {
  id: string
  title: string
  description: string
  image?: string
  images?: string[]
  longDescription?: string
  technologies?: string[]
  githubUrl?: string
  liveUrl?: string
  date?: string
  role?: string
  features?: string[]
  challenges?: string[]
  solutions?: string[]
}